Enable release, profile, and debug builds of vireo.js
