+++ package.json
@@ -24,7 +24,7 @@
     "xhr2": "^0.1.4"
   },
   "scripts": {
-    "build": "rollup source/core/vireo.loader.js --format umd --name \"vireoHelpers\" --file dist/wasm32-unknown-emscripten/release/vireo.umd.js",
+    "rollup": "cd make-it && rollup",
     "test": "karma start --browsers ChromeHeadlessNoSandbox,FirefoxHeadless",
     "test-dev": "karma start --browsers ChromeHeadlessNoSandbox --skip-tags Slow",
     "test-manual": "karma start --no-single-run",
@@ -50,6 +50,7 @@
     "url": "https://github.com/NI/VireoSDK/issues"
   },
   "homepage": "https://github.com/NI/VireoSDK#readme",
-  "main": "dist/wasm32-unknown-emscripten/release/vireo.umd.js",
+  "main": "dist/wasm32-unknown-emscripten/release/vireo.js",
+  "module": "source/core/vireo.loader.release.js",
   "dependencies": {}
 }
