+++ test-it/karma/static/VireoCodec.Test.js
@@ -1,7 +1,7 @@
 describe('Vireo', function () {
     'use strict';
 
-    var vireoHelpers = window.vireoHelpers;
+    var staticHelpers = window.vireoHelpers.staticHelpers;
 
     describe('can encode vireo identifiers', function () {
         describe('throws an exception when parameter', function () {
@@ -10,7 +10,7 @@ describe('Vireo', function () {
                     var obj = {
                         variable: 'to_encode'
                     };
-                    vireoHelpers.encodeIdentifier(obj);
+                    staticHelpers.encodeIdentifier(obj);
                 };
 
                 expect(encode).toThrow();
@@ -18,7 +18,7 @@ describe('Vireo', function () {
 
             it('is an empty string', function () {
                 var encode = function () {
-                    vireoHelpers.encodeIdentifier('');
+                    staticHelpers.encodeIdentifier('');
                 };
 
                 expect(encode).toThrow();
@@ -26,27 +26,27 @@ describe('Vireo', function () {
         });
 
         it('url encodes the first character if is not a letter [A-Za-z] and it is an ascii character', function () {
-            var encoded = vireoHelpers.encodeIdentifier('64bitInteger');
+            var encoded = staticHelpers.encodeIdentifier('64bitInteger');
             expect(encoded).toBe('%364bitInteger');
         });
 
         it('does not url-encode letters or numbers', function () {
-            var encoded = vireoHelpers.encodeIdentifier('ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789');
+            var encoded = staticHelpers.encodeIdentifier('ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789');
             expect(encoded).toBe('ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789');
         });
 
         it('does not url-encode symbols * + _ $ -', function () {
-            var encoded = vireoHelpers.encodeIdentifier('ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789*+_$-');
+            var encoded = staticHelpers.encodeIdentifier('ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789*+_$-');
             expect(encoded).toBe('ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789*+_$-');
         });
 
         it('does not encode non-ascii characters', function () {
-            var encoded = vireoHelpers.encodeIdentifier('Iñtërnâtiônàlizætiøn☃💩');
+            var encoded = staticHelpers.encodeIdentifier('Iñtërnâtiônàlizætiøn☃💩');
             expect(encoded).toBe('Iñtërnâtiônàlizætiøn☃💩');
         });
 
         it('url-encodes all other characters', function () {
-            var encoded = vireoHelpers.encodeIdentifier(' !"#%&\'(),./:;<=>?@[\\]^`{|}~');
+            var encoded = staticHelpers.encodeIdentifier(' !"#%&\'(),./:;<=>?@[\\]^`{|}~');
             expect(encoded).toBe('%20%21%22%23%25%26%27%28%29%2C%2E%2F%3A%3B%3C%3D%3E%3F%40%5B%5C%5D%5E%60%7B%7C%7D%7E');
         });
     });
@@ -58,7 +58,7 @@ describe('Vireo', function () {
                     var obj = {
                         variable: 'to_encode'
                     };
-                    vireoHelpers.decodeIdentifier(obj);
+                    staticHelpers.decodeIdentifier(obj);
                 };
 
                 expect(decode).toThrow();
@@ -66,7 +66,7 @@ describe('Vireo', function () {
 
             it('is an empty string', function () {
                 var decode = function () {
-                    vireoHelpers.decodeIdentifier('');
+                    staticHelpers.decodeIdentifier('');
                 };
 
                 expect(decode).toThrow();
@@ -74,17 +74,17 @@ describe('Vireo', function () {
         });
 
         it('does not decode non-ascii characters', function () {
-            var decoded = vireoHelpers.decodeIdentifier('Iñtërnâtiônàlizætiøn☃💩');
+            var decoded = staticHelpers.decodeIdentifier('Iñtërnâtiônàlizætiøn☃💩');
             expect(decoded).toBe('Iñtërnâtiônàlizætiøn☃💩');
         });
 
         it('does not decode symbols * + _ $ -', function () {
-            var decoded = vireoHelpers.decodeIdentifier('ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789*+_$-');
+            var decoded = staticHelpers.decodeIdentifier('ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789*+_$-');
             expect(decoded).toBe('ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789*+_$-');
         });
 
         it('decodes reserved characters', function () {
-            var decoded = vireoHelpers.decodeIdentifier('%20%21%22%23%25%26%27%28%29%2C%2E%2F%3A%3B%3C%3D%3E%3F%40%5B%5C%5D%5E%60%7B%7C%7D%7E');
+            var decoded = staticHelpers.decodeIdentifier('%20%21%22%23%25%26%27%28%29%2C%2E%2F%3A%3B%3C%3D%3E%3F%40%5B%5C%5D%5E%60%7B%7C%7D%7E');
             expect(decoded).toBe(' !"#%&\'(),./:;<=>?@[\\]^`{|}~');
         });
     });
